rename hazard runner modules
